Update Upay room and electricity integrations

+++ migrations/2025-10-28-025535-0000_create_rooms_table/up.sql
@@ -1,7 +1,7 @@
 -- 创建rooms表
 CREATE TABLE rooms (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    roomid INTEGER NOT NULL,
+    roomid BIGINT NOT NULL,
     electricity_fee REAL NOT NULL DEFAULT 0.0,
     send_flag BOOLEAN NOT NULL DEFAULT FALSE,
     threshold REAL NOT NULL,
