+++ migrations/2025-11-11-115447-0000_create_room_paths_table/up.sql
@@ -2,7 +2,7 @@
 
 CREATE TABLE room_paths (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    roomid INTEGER NOT NULL,
+    roomid BIGINT NOT NULL,
     
     roompath VARCHAR(255) NOT NULL UNIQUE,
     roompath_hash BIGINT NOT NULL,
